Enable affected tests
