+++ src/test/java/io/openliberty/tools/intellij/it/SingleModMPProjectTestCommon.java
@@ -145,6 +145,169 @@ public abstract class SingleModMPProjectTestCommon {
     }
 
     /**
+     * Tests Liberty tool window start.../stop actions selected on the project's drop-down action
+     * menu and run using the play action button on the Liberty tool window's toolbar.
+     */
+    @Test
+    @Video
+    public void testStartWithParamsActionUsingPlayToolbarButton() {
+        String testName = "testStartWithParamsActionUsingPlayToolbarButton";
+        String absoluteWLPPath = Paths.get(getProjectsDirPath(), getSmMPProjectName(), getWLPInstallPath()).toString();
+
+        // Delete any existing test report files.
+        deleteTestReports();
+
+        // Remove all other configurations first.
+        UIBotTestUtils.deleteLibertyRunConfigurations(remoteRobot);
+
+        // Trigger the start with parameters configuration dialog.
+        UIBotTestUtils.runLibertyActionFromLTWDropDownMenu(remoteRobot, "Start...", true);
+
+        // Run the configuration dialog.
+        UIBotTestUtils.runStartParamsConfigDialog(remoteRobot, getStartParams());
+
+        try {
+            // Validate that the application started.
+            TestUtils.validateProjectStarted(testName, getSmMpProjResURI(), getSmMpProjPort(), getSmMPProjOutput(), absoluteWLPPath, false);
+
+            // Validate that the report was generated.
+            validateTestReportsExist();
+        } finally {
+            if (TestUtils.isServerStopNeeded(absoluteWLPPath)) {
+                // Sleep for a few seconds to allow dev mode to finish running the tests. Specially
+                // for those times when the tests are run twice. Not waiting, opens up a window
+                // that leads to false negative results, and the Liberty server being left active.
+                // If the Liberty server is left active, subsequent tests will fail.
+                TestUtils.sleepAndIgnoreException(30);
+
+                // Stop Liberty dev mode and validates that the Liberty server is down.
+                UIBotTestUtils.runStopAction(remoteRobot, testName, UIBotTestUtils.ActionExecType.LTWPLAY, absoluteWLPPath, getSmMPProjectName(), 3);
+            }
+        }
+
+        // Validate that the start with params action brings up the configuration previously used.
+        try {
+            UIBotTestUtils.runLibertyActionFromLTWDropDownMenu(remoteRobot, "Start...", true);
+            Map<String, String> cfgEntries = UIBotTestUtils.getOpenedLibertyConfigDataAndCloseOnExit(remoteRobot);
+            String activeCfgName = cfgEntries.get(UIBotTestUtils.ConfigEntries.NAME.toString());
+            Assertions.assertEquals(getSmMPProjectName(), activeCfgName, "The active config name " + activeCfgName + " does not match expected name of " + getSmMPProjectName());
+            String activeCfgParams = cfgEntries.get(UIBotTestUtils.ConfigEntries.PARAMS.toString());
+            Assertions.assertEquals(getStartParams(), activeCfgParams, "The active config params " + activeCfgParams + " does not match expected params of " + getStartParams());
+        } finally {
+            // Cleanup configurations.
+            UIBotTestUtils.deleteLibertyRunConfigurations(remoteRobot);
+        }
+    }
+
+    /**
+     * Tests Liberty tool window start.../stop actions run from the project's pop-up action menu.
+     */
+    @Test
+    @Video
+    public void testStartWithParamsActionUsingPopUpMenu() {
+        String testName = "testStartWithParamsActionUsingPopUpMenu";
+        String absoluteWLPPath = Paths.get(getProjectsDirPath(), getSmMPProjectName(), getWLPInstallPath()).toString();
+
+        // Delete any existing test report files.
+        deleteTestReports();
+
+        // Remove all other configurations first.
+        UIBotTestUtils.deleteLibertyRunConfigurations(remoteRobot);
+
+        // Trigger the start with parameters configuration dialog.
+        UIBotTestUtils.runActionLTWPopupMenu(remoteRobot, getSmMPProjectName(), "Liberty: Start...");
+
+        // Run the configuration dialog.
+        UIBotTestUtils.runStartParamsConfigDialog(remoteRobot, getStartParams());
+
+        try {
+            // Validate that the application started.
+            TestUtils.validateProjectStarted(testName, getSmMpProjResURI(), getSmMpProjPort(), getSmMPProjOutput(), absoluteWLPPath, false);
+
+            // Validate that the report was generated.
+            validateTestReportsExist();
+        } finally {
+            if (TestUtils.isServerStopNeeded(absoluteWLPPath)) {
+                // Sleep for a few seconds to allow dev mode to finish running the tests. Specially
+                // for those times when the tests are run twice. Not waiting, opens up a window
+                // that leads to false negative results, and the Liberty server being left active.
+                // If the Liberty server is left active, subsequent tests will fail.
+                TestUtils.sleepAndIgnoreException(30);
+
+                // Stop Liberty dev mode and validates that the Liberty server is down.
+                UIBotTestUtils.runStopAction(remoteRobot, testName, UIBotTestUtils.ActionExecType.LTWPOPUP, absoluteWLPPath, getSmMPProjectName(), 3);
+            }
+        }
+
+        // Validate that the start with params action brings up the configuration previously used.
+        try {
+            UIBotTestUtils.runActionLTWPopupMenu(remoteRobot, getSmMPProjectName(), "Liberty: Start...");
+            Map<String, String> cfgEntries = UIBotTestUtils.getOpenedLibertyConfigDataAndCloseOnExit(remoteRobot);
+            String activeCfgName = cfgEntries.get(UIBotTestUtils.ConfigEntries.NAME.toString());
+            Assertions.assertEquals(getSmMPProjectName(), activeCfgName, "The active config name " + activeCfgName + " does not match expected name of " + getSmMPProjectName());
+            String activeCfgParams = cfgEntries.get(UIBotTestUtils.ConfigEntries.PARAMS.toString());
+            Assertions.assertEquals(getStartParams(), activeCfgParams, "The active config params " + activeCfgParams + " does not match expected params of " + getStartParams());
+        } finally {
+            // Cleanup configurations.
+            UIBotTestUtils.deleteLibertyRunConfigurations(remoteRobot);
+        }
+    }
+
+    /**
+     * Tests start.../stop actions run from the search everywhere panel.
+     */
+    @Test
+    @Video
+    public void testStartWithParamsActionUsingSearch() {
+        String testName = "testStartWithParamsActionUsingSearch";
+        String absoluteWLPPath = Paths.get(getProjectsDirPath(), getSmMPProjectName(), getWLPInstallPath()).toString();
+
+        // Delete any existing test report files.
+        deleteTestReports();
+
+        // Remove all other configurations first.
+        UIBotTestUtils.deleteLibertyRunConfigurations(remoteRobot);
+
+        // Trigger the start with parameters configuration dialog.
+        UIBotTestUtils.runActionFromSearchEverywherePanel(remoteRobot, "Liberty: Start...");
+
+        // Run the configuration dialog.
+        UIBotTestUtils.runStartParamsConfigDialog(remoteRobot, getStartParams());
+
+        try {
+            // Validate that the application started.
+            TestUtils.validateProjectStarted(testName, getSmMpProjResURI(), getSmMpProjPort(), getSmMPProjOutput(), absoluteWLPPath, false);
+
+            // Validate that the report was generated.
+            validateTestReportsExist();
+        } finally {
+            if (TestUtils.isServerStopNeeded(absoluteWLPPath)) {
+                // Sleep for a few seconds to allow dev mode to finish running the tests. Specially
+                // for those times when the tests are run twice. Not waiting, opens up a window
+                // that leads to false negative results, and the Liberty server being left active.
+                // If the Liberty server is left active, subsequent tests will fail.
+                TestUtils.sleepAndIgnoreException(30);
+
+                // Stop Liberty dev mode and validates that the Liberty server is down.
+                UIBotTestUtils.runStopAction(remoteRobot, testName, UIBotTestUtils.ActionExecType.SEARCH, absoluteWLPPath, getSmMPProjectName(), 3);
+            }
+        }
+
+        // Validate that the start with params action brings up the configuration previously used.
+        try {
+            UIBotTestUtils.runActionFromSearchEverywherePanel(remoteRobot, "Liberty: Start...");
+            Map<String, String> cfgEntries = UIBotTestUtils.getOpenedLibertyConfigDataAndCloseOnExit(remoteRobot);
+            String activeCfgName = cfgEntries.get(UIBotTestUtils.ConfigEntries.NAME.toString());
+            Assertions.assertEquals(getSmMPProjectName(), activeCfgName, "The active config name " + activeCfgName + " does not match expected name of " + getSmMPProjectName());
+            String activeCfgParams = cfgEntries.get(UIBotTestUtils.ConfigEntries.PARAMS.toString());
+            Assertions.assertEquals(getStartParams(), activeCfgParams, "The active config params " + activeCfgParams + " does not match expected params of " + getStartParams());
+        } finally {
+            // Cleanup configurations.
+            UIBotTestUtils.deleteLibertyRunConfigurations(remoteRobot);
+        }
+    }
+
+    /**
      * Tests dashboard start/RunTests/stop actions run from the project's drop-down action menu.
      */
     @Test
@@ -170,11 +333,129 @@ public abstract class SingleModMPProjectTestCommon {
             validateTestReportsExist();
         } finally {
             if (TestUtils.isServerStopNeeded(absoluteWLPPath)) {
-                // Stop dev mode.
-                UIBotTestUtils.runLibertyActionFromLTWDropDownMenu(remoteRobot, "Stop", false);
+                // Sleep for a few seconds to allow dev mode to finish running the tests. Specially
+                // for those times when the tests are run twice. Not waiting, opens up a window
+                // that leads to false negative results, and the Liberty server being left active.
+                // If the Liberty server is left active, subsequent tests will fail.
+                TestUtils.sleepAndIgnoreException(30);
 
-                // Validate that the server stopped.
-                TestUtils.validateLibertyServerStopped(testName, absoluteWLPPath);
+                // Stop Liberty dev mode and validates that the Liberty server is down.
+                UIBotTestUtils.runStopAction(remoteRobot, testName, UIBotTestUtils.ActionExecType.LTWDROPDOWN, absoluteWLPPath, getSmMPProjectName(), 3);
+            }
+        }
+    }
+
+    /**
+     * Tests Liberty tool window start/RunTests/stop actions selected on the project's drop-down action
+     * menu and run using the play action button on the Liberty tool window's toolbar.
+     */
+    @Test
+    @Video
+    public void testRunTestsActionUsingPlayToolbarButton() {
+        String testName = "testRunTestsActionUsingPlayToolbarButton";
+        String absoluteWLPPath = Paths.get(getProjectsDirPath(), getSmMPProjectName(), getWLPInstallPath()).toString();
+
+        // Delete any existing test report files.
+        deleteTestReports();
+
+        // Start dev mode.
+        UIBotTestUtils.runLibertyActionFromLTWDropDownMenu(remoteRobot, "Start", true);
+
+        try {
+            // Validate that the application started.
+            TestUtils.validateProjectStarted(testName, getSmMpProjResURI(), getSmMpProjPort(), getSmMPProjOutput(), absoluteWLPPath, false);
+
+            // Run the application's tests.
+            UIBotTestUtils.runLibertyActionFromLTWDropDownMenu(remoteRobot, "Run tests", true);
+
+            // Validate that the report was generated.
+            validateTestReportsExist();
+        } finally {
+            if (TestUtils.isServerStopNeeded(absoluteWLPPath)) {
+                // Sleep for a few seconds to allow dev mode to finish running the tests. Specially
+                // for those times when the tests are run twice. Not waiting, opens up a window
+                // that leads to false negative results, and the Liberty server being left active.
+                // If the Liberty server is left active, subsequent tests will fail.
+                TestUtils.sleepAndIgnoreException(30);
+
+                // Stop Liberty dev mode and validates that the Liberty server is down.
+                UIBotTestUtils.runStopAction(remoteRobot, testName, UIBotTestUtils.ActionExecType.LTWDROPDOWN, absoluteWLPPath, getSmMPProjectName(), 3);
+            }
+        }
+    }
+
+    /**
+     * Tests Liberty tool window start/runTests/stop actions run from the project's pop-up action menu.
+     */
+    @Test
+    @Video
+    public void testRunTestsActionUsingPopUpMenu() {
+        String testName = "testRunTestsActionUsingPopUpMenu";
+        String absoluteWLPPath = Paths.get(getProjectsDirPath(), getSmMPProjectName(), getWLPInstallPath()).toString();
+
+        // Delete any existing test report files.
+        deleteTestReports();
+
+        // Start dev mode.
+        UIBotTestUtils.runActionLTWPopupMenu(remoteRobot, getSmMPProjectName(), "Liberty: Start");
+
+        try {
+            // Validate that the application started.
+            TestUtils.validateProjectStarted(testName, getSmMpProjResURI(), getSmMpProjPort(), getSmMPProjOutput(), absoluteWLPPath, false);
+
+            // Run the application's tests.
+            UIBotTestUtils.runActionLTWPopupMenu(remoteRobot, getSmMPProjectName(), "Liberty: Run tests");
+
+            // Validate that the reports were generated.
+            validateTestReportsExist();
+        } finally {
+            if (TestUtils.isServerStopNeeded(absoluteWLPPath)) {
+                // Sleep for a few seconds to allow dev mode to finish running the tests. Specially
+                // for those times when the tests are run twice. Not waiting, opens up a window
+                // that leads to false negative results, and the Liberty server being left active.
+                // If the Liberty server is left active, subsequent tests will fail.
+                TestUtils.sleepAndIgnoreException(30);
+
+                // Stop Liberty dev mode and validates that the Liberty server is down.
+                UIBotTestUtils.runStopAction(remoteRobot, testName, UIBotTestUtils.ActionExecType.LTWPOPUP, absoluteWLPPath, getSmMPProjectName(), 3);
+            }
+        }
+    }
+
+    /**
+     * Tests start/runTests/stop actions run from the search everywhere panel.
+     */
+    @Test
+    @Video
+    public void testRunTestsActionUsingSearch() {
+        String testName = "testRunTestsActionUsingSearch";
+        String absoluteWLPPath = Paths.get(getProjectsDirPath(), getSmMPProjectName(), getWLPInstallPath()).toString();
+
+        // Delete any existing test report files.
+        deleteTestReports();
+
+        // Start dev mode.
+        UIBotTestUtils.runActionFromSearchEverywherePanel(remoteRobot, "Liberty: Start");
+
+        try {
+            // Validate that the application started.
+            TestUtils.validateProjectStarted(testName, getSmMpProjResURI(), getSmMpProjPort(), getSmMPProjOutput(), absoluteWLPPath, false);
+
+            // Run the application's tests.
+            UIBotTestUtils.runActionFromSearchEverywherePanel(remoteRobot, "Liberty: Run tests");
+
+            // Validate that the reports were generated.
+            validateTestReportsExist();
+        } finally {
+            if (TestUtils.isServerStopNeeded(absoluteWLPPath)) {
+                // Sleep for a few seconds to allow dev mode to finish running the tests. Specially
+                // for those times when the tests are run twice. Not waiting, opens up a window
+                // that leads to false negative results, and the Liberty server being left active.
+                // If the Liberty server is left active, subsequent tests will fail.
+                TestUtils.sleepAndIgnoreException(30);
+
+                // Stop Liberty dev mode and validates that the Liberty server is down.
+                UIBotTestUtils.runStopAction(remoteRobot, testName, UIBotTestUtils.ActionExecType.SEARCH, absoluteWLPPath, getSmMPProjectName(), 3);
             }
         }
     }
